JAVA-1364: Enable creation of SslHandler with remote address information
Motivation:

SSLOptions do not currently allow to create SslHandler instances with remote address information.
This is needed to instantiate SSLEngine/SslHandler with remote address information set,
which enables doing hostname verification (for example through a custom TrustManager).
We cannot take this information from SocketChannel since it is not connected yet
at the time Connection.Initializer creates the SslHandler.

Modifications:

Create a child interface to SSLOptions, EndpointAwareSSLOptions, that exposes a new
method to create SslHandler instances with remote address information.
Modify Connection class so that it detects instances of EndpointAwareSSLOptions
and calls the new method instead of the old one.
Modify current implementations of SSLOptions, JdkSSLOptions and NettySSLOptions
to implement the new interface; delegate old methods to new ones.

Result:

Remote endpoint information is now available to implementors of EndpointAwareSSLOptions,
and to the driver built-in implementations JdkSSLOptions and NettySSLOptions.

This commit has substantial contributions from Tim Lamballais Tessensohn (@wimtie).